+++ truss/remote/baseten/api.py
@@ -570,19 +570,15 @@ class BasetenApi:
         resp_json = self._rest_api_client.get(f"v1/training_projects/{project_id}/jobs")
         return resp_json
 
-    def list_all_training_jobs(
+    def search_training_jobs(
         self,
-        status_filter: Optional[List[str]] = None,
+        statuses: Optional[List[str]] = None,
         project_id: Optional[str] = None,
         job_id: Optional[str] = None,
     ):
         resp_json = self._rest_api_client.post(
-            "v1/training_projects/all_jobs",
-            body={
-                "status_filter": status_filter,
-                "project_id": project_id,
-                "job_id": job_id,
-            },
+            "v1/training_jobs/search",
+            body={"statuses": statuses, "project_id": project_id, "job_id": job_id},
         )
         return resp_json["training_jobs"]
